Convert font to outlines for sharing SVG version of hex sticker.
The main SVG image requires that the font Myriad Pro Semi-Bold
be installed on the machine. This makes it hard to share because
it looks bad w/o that font.

This isn't a problem for the rasterized PNG version, but I want
to share the vector format SVG for the rstudio::conf hex sticker
wall.

To fix this, I needed to convert the text to outlines. This was
suprisingly hard to find the solution online, but it's really
easy, as shown in this video:

https://www.youtube.com/watch?v=_01Jdvr7eXI

Path -> Object to Path